use emotion
