+++ src/plugins/charts/common/constants.ts
@@ -20,3 +20,23 @@ export const paletteIds = [
   'warm',
   'gray',
 ];
+
+// This set of defaults originated in Canvas, which, at present, is the primary
+// consumer of this function.  Changing this default requires a change in Canvas
+// logic, which would likely be a breaking change in 7.x.
+export const defaultCustomColors = [
+  '#882E72',
+  '#B178A6',
+  '#D6C1DE',
+  '#1965B0',
+  '#5289C7',
+  '#7BAFDE',
+  '#4EB265',
+  '#90C987',
+  '#CAE0AB',
+  '#F7EE55',
+  '#F6C141',
+  '#F1932D',
+  '#E8601C',
+  '#DC050C',
+];
