--- src/plugins/charts/public/static/components/color_picker.scss
@@ -1,30 +1,0 @@
-$visColorPickerWidth: $euiSizeL * 8; // 8 columns
-
-.visColorPicker__value {
-  width: $visColorPickerWidth;
-}
-
-.visColorPicker__colorBtn {
-  position: relative;
-
-  input[type='radio'] {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    opacity: 0;
-    transform: translate(-50%, -50%);
-  }
-}
-
-.visColorPicker__valueDot {
-  cursor: pointer;
-
-  &:hover {
-    transform: scale(1.4);
-  }
-
-  &-isSelected {
-    border: $euiSizeXS solid;
-    border-radius: 100%;
-  }
-}
