--- src/plugins/charts/public/static/components/common_chart_styles.scss
@@ -1,7 +1,0 @@
-.chart-icon__subdued {
-  fill: $euiTextSubduedColor;
-}
-
-.chart-icon__accent {
-  fill: $euiColorVis0;
-}
