--- src/plugins/charts/public/static/components/empty_placeholder.scss
@@ -1,7 +1,0 @@
-.chart__empty-placeholder {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
