--- src/plugins/charts/public/static/components/legend_toggle.scss
@@ -1,19 +1,0 @@
-.echLegend__toggle {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  z-index: 1;
-  margin: $euiSizeXS;
-
-  &--isOpen {
-    background-color: $euiColorLightestShade;
-  }
-
-  &--position-left,
-  &--position-bottom {
-    left: auto;
-    bottom: auto;
-    right: 0;
-    top: 0;
-  }
-}
