+++ src/plugins/charts/public/static/components/legend_toggle.tsx
@@ -7,13 +7,10 @@
  */
 
 import React, { memo, useMemo } from 'react';
-import classNames from 'classnames';
-
 import { i18n } from '@kbn/i18n';
-import { htmlIdGenerator, EuiButtonIcon } from '@elastic/eui';
+import { htmlIdGenerator, EuiButtonIcon, useEuiTheme } from '@elastic/eui';
 import { Position } from '@elastic/charts';
-
-import './legend_toggle.scss';
+import { css } from '@emotion/react';
 
 interface LegendToggleProps {
   onClick: () => void;
@@ -23,6 +20,35 @@ interface LegendToggleProps {
 
 const LegendToggleComponent = ({ onClick, showLegend, legendPosition }: LegendToggleProps) => {
   const legendId = useMemo(() => htmlIdGenerator()('legend'), []);
+  const { euiTheme } = useEuiTheme();
+
+  const baseStyles = useMemo(
+    () => css`
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      z-index: 1;
+      margin: ${euiTheme.size.xs};
+    `,
+    [euiTheme.size.xs]
+  );
+
+  const isOpenStyle = useMemo(
+    () => css`
+      background-color: ${euiTheme.colors.lightestShade};
+    `,
+    [euiTheme.colors.lightestShade]
+  );
+
+  const positionStyle = useMemo(
+    () => css`
+      left: auto;
+      bottom: auto;
+      right: 0;
+      top: 0;
+    `,
+    []
+  );
 
   return (
     <EuiButtonIcon
@@ -30,9 +56,11 @@ const LegendToggleComponent = ({ onClick, showLegend, legendPosition }: LegendTo
       iconType="list"
       color="text"
       onClick={onClick}
-      className={classNames('echLegend__toggle', `echLegend__toggle--position-${legendPosition}`, {
-        'echLegend__toggle--isOpen': showLegend,
-      })}
+      css={[
+        baseStyles,
+        showLegend ? isOpenStyle : null,
+        ['left', 'bottom'].includes(legendPosition) ? positionStyle : null,
+      ]}
       aria-label={i18n.translate('charts.legend.toggleLegendButtonAriaLabel', {
         defaultMessage: 'Toggle legend',
       })}
